Use double to avoid precision problem.

+++ AppSettings.cs
@@ -37,7 +37,7 @@ public class AppSettings
     [JsonIgnore]
     public string Language => Locale.Length > 2 ? Locale[..2] : Locale;
 
-    public float VoiceConfidence { get; set; } = 0.6f;
+    public double VoiceConfidence { get; set; } = 0.6f;
     public string WakeupWord { get; set; } = "";
     public bool EnableHotkeyTrigger { get; set; } = true;
     public string UpdateUrl { get; set; } = "https://github.com/tifish/HellDivers2OneKeyStratagem/releases/download/latest_release/HellDivers2OneKeyStratagem.7z";
